refactor: extract TripMetadataBuilder from TripDownloadService
Extract entity building logic (areas, places, segments, polygons) into
a dedicated ITripMetadataBuilder service. This eliminates code
duplication between DownloadTripAsync and SyncTripAsync methods.

- Add ITripMetadataBuilder interface with pure transformation methods
- Add TripMetadataBuilder implementation
- Update TripDownloadService to use the new service
- Register in DI container

Part of issue #77 (reduce service sizes).
