+++ src/WayfarerMobile/MauiProgram.cs
@@ -182,6 +182,7 @@ public static class MauiProgram
         services.AddSingleton<ITileDownloadService, TileDownloadService>();
         services.AddSingleton<IDownloadStateManager, DownloadStateManager>();
         services.AddSingleton<ICacheLimitEnforcer, CacheLimitEnforcer>();
+        services.AddSingleton<ITripMetadataBuilder, TripMetadataBuilder>();
         services.AddSingleton<TripDownloadService>();
         // Also register as interface for consumers that prefer interface injection
         services.AddSingleton<ITripDownloadService>(sp => sp.GetRequiredService<TripDownloadService>());
